formatting

+++ src/SampleDocker/GetStorageItemsEndpoint.cs
@@ -2,7 +2,6 @@
 using Azure.Storage.Blobs;
 using FatCat.Toolkit.Console;
 using FatCat.Toolkit.Extensions;
-using FatCat.Toolkit.Web;
 using FatCat.Toolkit.WebServer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -16,10 +15,7 @@ public class GetStorageItemsEndpoint : Endpoint
 	private const string ContainerName = "quickstartblobs-da3c108f-fb28-47cb-8e57-4a2578253c3c";
 	private readonly IConfiguration configuration;
 
-	public GetStorageItemsEndpoint(IConfiguration configuration)
-	{
-		this.configuration = configuration;
-	}
+	public GetStorageItemsEndpoint(IConfiguration configuration) => this.configuration = configuration;
 
 	[HttpPost("api/store")]
 	public async Task<WebResult> GetStorageItems()
@@ -50,21 +46,12 @@ public class GetStorageItemsEndpoint : Endpoint
 		// Create the container and return a container client object
 		var items = blobServiceClient.GetBlobContainers();
 
-		foreach (var item in items)
-		{
-			ConsoleLog.WriteCyan($"{item.Name}");
-		}
+		foreach (var item in items) { ConsoleLog.WriteCyan($"{item.Name}"); }
 
 		BlobContainerClient containerClient;
 
-		if (items.All(i => i.Name != ContainerName))
-		{
-			containerClient = await blobServiceClient.CreateBlobContainerAsync(ContainerName);
-		}
-		else
-		{
-			containerClient = blobServiceClient.GetBlobContainerClient(ContainerName);
-		}
+		if (items.All(i => i.Name != ContainerName)) { containerClient = await blobServiceClient.CreateBlobContainerAsync(ContainerName); }
+		else { containerClient = blobServiceClient.GetBlobContainerClient(ContainerName); }
 
 		// Create a local file in the ./data/ directory for uploading and downloading
 		var localPath = "data";
@@ -72,10 +59,7 @@ public class GetStorageItemsEndpoint : Endpoint
 		var fileName = "quickstart_Dude.txt";
 		var localFilePath = Path.Combine(localPath, fileName);
 
-		if (System.IO.File.Exists(localFilePath))
-		{
-			System.IO.File.Delete(localFilePath);
-		}
+		if (System.IO.File.Exists(localFilePath)) { System.IO.File.Delete(localFilePath); }
 
 		// Write text to the file
 		await System.IO.File.WriteAllTextAsync(localFilePath, $"Hello, World! | <{DateTime.Now:hh:mm:ss}>");
