+++ src/ToolKit/FileSystemTools.cs
@@ -45,10 +45,7 @@ public class FileSystemTools : IFileSystemTools
 {
 	private readonly IFileSystem fileSystem;
 
-	public FileSystemTools(IFileSystem fileSystem)
-	{
-		this.fileSystem = fileSystem;
-	}
+	public FileSystemTools(IFileSystem fileSystem) => this.fileSystem = fileSystem;
 
 	public async Task AppendToFile(string path, string text)
 	{
@@ -59,35 +56,23 @@ public class FileSystemTools : IFileSystemTools
 
 	public void DeleteDirectory(string path, bool recursive = true)
 	{
-		if (DirectoryExists(path))
-		{
-			fileSystem.Directory.Delete(path, recursive);
-		}
+		if (DirectoryExists(path)) { fileSystem.Directory.Delete(path, recursive); }
 	}
 
 	public bool DeleteFile(string path)
 	{
-		if (!FileExists(path))
-		{
-			return false;
-		}
+		if (!FileExists(path)) { return false; }
 
 		fileSystem.File.Delete(path);
 
 		return true;
 	}
 
-	public bool DirectoryExists(string path)
-	{
-		return fileSystem.Directory.Exists(path);
-	}
+	public bool DirectoryExists(string path) => fileSystem.Directory.Exists(path);
 
 	public void EnsureDirectory(string path)
 	{
-		if (DirectoryExists(path))
-		{
-			return;
-		}
+		if (DirectoryExists(path)) { return; }
 
 		fileSystem.Directory.CreateDirectory(path);
 	}
@@ -96,42 +81,27 @@ public class FileSystemTools : IFileSystemTools
 	{
 		EnsureDirectory(Path.GetDirectoryName(path)!);
 
-		if (FileExists(path))
-		{
-			return;
-		}
+		if (FileExists(path)) { return; }
 
 		using var _ = fileSystem.File.Create(path);
 	}
 
-	public bool FileExists(string path)
-	{
-		return fileSystem.File.Exists(path);
-	}
+	public bool FileExists(string path) => fileSystem.File.Exists(path);
 
 	public List<string> GetDirectories(string path)
 	{
-		if (!DirectoryExists(path))
-		{
-			return new();
-		}
+		if (!DirectoryExists(path)) { return new(); }
 
 		var directories = fileSystem.Directory.GetDirectories(path);
 
 		return directories.ToList();
 	}
 
-	public IFileInfo GetFileMetaData(string fullPath)
-	{
-		return fileSystem.FileInfo.New(fullPath);
-	}
+	public IFileInfo GetFileMetaData(string fullPath) => fileSystem.FileInfo.New(fullPath);
 
 	public List<string> GetFiles(string directoryPath)
 	{
-		if (!DirectoryExists(directoryPath))
-		{
-			return new();
-		}
+		if (!DirectoryExists(directoryPath)) { return new(); }
 
 		var files = fileSystem.Directory.GetFiles(directoryPath);
 
@@ -147,10 +117,7 @@ public class FileSystemTools : IFileSystemTools
 
 	public bool MoveDirectory(string sourceDirectory, string destinationDirectory)
 	{
-		if (!DirectoryExists(sourceDirectory))
-		{
-			return false;
-		}
+		if (!DirectoryExists(sourceDirectory)) { return false; }
 
 		fileSystem.Directory.Move(sourceDirectory, destinationDirectory);
 
@@ -159,37 +126,25 @@ public class FileSystemTools : IFileSystemTools
 
 	public bool MoveFile(string sourcePath, string sourceDestination)
 	{
-		if (!FileExists(sourcePath))
-		{
-			return false;
-		}
+		if (!FileExists(sourcePath)) { return false; }
 
 		fileSystem.File.Move(sourcePath, sourceDestination);
 
 		return true;
 	}
 
-	public async Task<byte[]> ReadAllBytes(string path)
-	{
-		return FileDoesNotExist(path) ? Array.Empty<byte>() : await fileSystem.File.ReadAllBytesAsync(path);
-	}
+	public async Task<byte[]> ReadAllBytes(string path) => FileDoesNotExist(path) ? Array.Empty<byte>() : await fileSystem.File.ReadAllBytesAsync(path);
 
 	public async Task<List<string>> ReadAllLines(string path)
 	{
-		if (!fileSystem.File.Exists(path))
-		{
-			return new List<string>();
-		}
+		if (!fileSystem.File.Exists(path)) { return new List<string>(); }
 
 		var lines = await fileSystem.File.ReadAllLinesAsync(path);
 
 		return lines.ToList();
 	}
 
-	public async Task<string> ReadAllText(string path)
-	{
-		return FileDoesNotExist(path) ? string.Empty : await fileSystem.File.ReadAllTextAsync(path);
-	}
+	public async Task<string> ReadAllText(string path) => FileDoesNotExist(path) ? string.Empty : await fileSystem.File.ReadAllTextAsync(path);
 
 	public async Task WriteAllBytes(string path, byte[] bytes)
 	{
@@ -205,8 +160,5 @@ public class FileSystemTools : IFileSystemTools
 		await fileSystem.File.WriteAllTextAsync(path, text);
 	}
 
-	private bool FileDoesNotExist(string path)
-	{
-		return !FileExists(path);
-	}
+	private bool FileDoesNotExist(string path) => !FileExists(path);
 }
